web/requirements.txt
aiodns==3.0.0
argh==0.26.2
beautifulsoup4==4.9.3
celery==5.4.0
discord-webhook==1.3.0
Django==3.2.23
django-ace==1.0.11
django-celery-beat==2.6.0
django-login-required-middleware==0.6.1
django-role-permissions==3.2.0
django-mathfilters==1.0.0
django-timezone-field==6.1.0
djangorestframework==3.12.4
djangorestframework-datatables==0.6.0
dotted-dict==1.1.3
drf-yasg==1.21.3
gunicorn==22.0.0
gevent==24.2.1
humanize==4.3.0
langchain-community==0.2.7
Markdown==3.3.4
metafinder==1.2
netaddr==0.8.0
netlas==0.5.1
openai==0.28.0
PyYAML==6.0.1
PySocks==1.7.1
psycopg2==2.9.7
pycvesearch==1.0
redis==5.0.3
requests==2.31.0
scapy==2.4.3
sqlalchemy==1.4.52
tldextract==3.5.0
uro==1.0.0
validators==0.18.2
watchdog==4.0.0
whatportis
weasyprint==53.3
wafw00f==2.2.0
xmltodict==0.13.0
django-environ==0.11.2
minify_html==0.15.0
numpy==1.26.4
transformers==4.40.2
torch==2.4.1
pika==1.3.2
django-cors-headers==4.3.0
django-cors-middleware==1.5.0